All postprocessing and all tools now built with autotools
This is the fist full pass at all FRE NCtools to be built using GNU autotools.  This will build all serial tools.  Work still needs to be done for the parallel tools.
