+++ tools/check_mask/Makefile.am
@@ -1,0 +1,8 @@
+bin_PROGRAMS = check_mask
+
+AM_CFLAGS = -I$(top_srcdir)/tools/libfrencutils \
+            $(NETCDF_CFLAGS)
+LDADD = $(top_builddir)/tools/libfrencutils/libfrencutils.a \
+        $(NETCDF_LDFLAGS) $(NETCDF_LIBS)
+
+check_mask_SOURCES = check_mask.c
