+++ tools/cubic_utils/Makefile.am
@@ -1,0 +1,8 @@
+bin_PROGRAMS = make_remap_file
+
+AM_CFLAGS = -I$(top_srcdir)/tools/libfrencutils \
+            $(NETCDF_CFLAGS)
+LDADD = $(top_builddir)/tools/libfrencutils/libfrencutils.a \
+        $(NETCDF_LDFLAGS) $(NETCDF_LIBS)
+
+make_remap_file_SOURCES = make_remap_file.c
