+++ tools/libfrencutils/COPYING
