+++ tools/libfrencutils/Makefile.am
@@ -1,0 +1,14 @@
+noinst_LIBRARIES = libfrencutils.a
+
+AM_CFLAGS = $(NETCDF_CFLAGS)
+
+libfrencutils_a_SOURCES = create_xgrid.c \
+                      gradient_c2l.c \
+                      interp.c \
+                      mosaic_util.c \
+                      read_mosaic.c \
+                      affinity.c \
+                      mpp_domain.c \
+                      mpp_io.c \
+                      mpp.c \
+                      tool_util.c
