+++ tools/libfrencutils/constant.h
