+++ tools/libfrencutils/gradient.F90
