+++ tools/libfrencutils/interp.c
