+++ tools/libfrencutils/mosaic_util.c
