+++ tools/libfrencutils/mpp.c
