+++ tools/libfrencutils/mpp_domain.c
