+++ tools/libfrencutils/mpp_io.c
