+++ tools/libfrencutils/read_mosaic.h
