+++ tools/libfrencutils/tool_util.c
