+++ tools/libfrencutils/tool_util.h
