+++ tools/make_coupler_mosaic/Makefile.am
@@ -1,0 +1,8 @@
+bin_PROGRAMS = make_coupler_mosaic
+
+AM_CFLAGS = -I$(top_srcdir)/tools/libfrencutils \
+            $(NETCDF_CFLAGS)
+LDADD = $(top_builddir)/tools/libfrencutils/libfrencutils.a \
+        $(NETCDF_LDFLAGS) $(NETCDF_LIBS)
+
+make_coupler_mosaic_SOURCES = make_coupler_mosaic.c
