+++ tools/make_land_domain/Makefile.am
@@ -1,0 +1,8 @@
+bin_PROGRAMS = make_land_domain
+
+AM_CFLAGS = -I$(top_srcdir)/tools/libfrencutils \
+            $(NETCDF_CFLAGS)
+LDADD = $(top_builddir)/tools/libfrencutils/libfrencutils.a \
+        $(NETCDF_LDFLAGS) $(NETCDF_LIBS)
+
+make_land_domain_SOURCES = make_land_domain.c
