+++ tools/make_quick_mosaic/Makefile.am
@@ -1,0 +1,8 @@
+bin_PROGRAMS = make_quick_mosaic
+
+AM_CFLAGS = -I$(top_srcdir)/tools/libfrencutils \
+            $(NETCDF_CFLAGS)
+LDADD = $(top_builddir)/tools/libfrencutils/libfrencutils.a \
+        $(NETCDF_LDFLAGS) $(NETCDF_LIBS)
+
+make_quick_mosaic_SOURCES = make_quick_mosaic.c
