+++ tools/make_vgrid/Makefile.am
@@ -1,0 +1,9 @@
+bin_PROGRAMS = make_vgrid
+
+AM_CFLAGS = -I$(top_srcdir)/tools/libfrencutils \
+            $(NETCDF_CFLAGS)
+LDADD = $(top_builddir)/tools/libfrencutils/libfrencutils.a \
+        $(NETCDF_LDFLAGS) $(NETCDF_LIBS)
+
+make_vgrid_SOURCES = create_vgrid.c \
+                     make_vgrid.c
