+++ tools/remap_land/Makefile.am
@@ -1,0 +1,8 @@
+bin_PROGRAMS = remap_land
+
+AM_CFLAGS = -I$(top_srcdir)/tools/libfrencutils \
+            $(NETCDF_CFLAGS)
+LDADD = $(top_builddir)/tools/libfrencutils/libfrencutils.a \
+        $(NETCDF_LDFLAGS) $(NETCDF_LIBS)
+
+remap_land_SOURCES = remap_land.c
