+++ tools/river_regrid/Makefile.am
@@ -1,0 +1,8 @@
+bin_PROGRAMS = river_regrid
+
+AM_CFLAGS = -I$(top_srcdir)/tools/libfrencutils \
+            $(NETCDF_CFLAGS)
+LDADD = $(top_builddir)/tools/libfrencutils/libfrencutils.a \
+        $(NETCDF_LDFLAGS) $(NETCDF_LIBS)
+
+river_regrid_SOURCES = river_regrid.c
