+++ tools/runoff_regrid/Makefile.am
@@ -1,0 +1,8 @@
+bin_PROGRAMS = runoff_regrid
+
+AM_CFLAGS = -I$(top_srcdir)/tools/libfrencutils \
+            $(NETCDF_CFLAGS)
+LDADD = $(top_builddir)/tools/libfrencutils/libfrencutils.a \
+        $(NETCDF_LDFLAGS) $(NETCDF_LIBS)
+
+runoff_regrid_SOURCES = runoff_regrid.c
